Changes to support both local and CRAN packages
Previoulsy a r.yml file like this wouldn't work:

```yml
---
packages:
- cran_mirror: https://cran.r-project.org
  packages:
      - name: shiny
      - name: ./vendor_r/RcppEigen.tar.gz
```

Problem
-------
Due to the fact that the system would reject paths containing dots, and that the `supply.go` file set a mirror name.

The problem is that users may want to "inject" a specfic package (perhaps it's internal or perhaps the package needs to be pre-compiled). This wasn't possible before because it was all or nothing. Either you used a local CRAN repository or you used a online mirror.

Solution
--------

The changes made in this patch:

1. If a non alpha or dot is found, `install.packages` mirror is set to NULL and the package is installed as a local package. Otherwise CRAN is used.
2. Each package is installed seperately, not combined into one `install.package`. This was done because often times the order was important, you may need two online packages, a local, then more online pacakages. For those local packages, they may have dependancies that you have to install first.

Cavets
------

This patch removed the local CRAN functionality, mostly for time when quickly implementing it.

Future changes
--------------

Likely the syntax and way this is done should be changed. Perhaps multiple package sections could be supported instead which would be more natural.